Fix for 64 bit compatibility.
- Use %cmake macro to set default path
- Fix hard-coded library path
- Fix the wrong cast between int and pointer.

+++ src/client.cpp
@@ -556,7 +556,7 @@ void lcd_off_set_wake_up(keynode_t *node, void *data)
 
 static void sensor_changed_cb(keynode_t *node, void *data)
 {
-	int event_number = (int)(data);
+	signed long event_number = (signed long)(data);
 	unsigned int i = 0;
 	int val;
 	int cb_number = 0;
